refactor: split build.ts into multiple files with one step in each
Towards #1546
